% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{add.xaxis.to.perm.plot}
\alias{add.xaxis.to.perm.plot}
\title{X-axis for permutation plot}
\usage{
add.xaxis.to.perm.plot(
  xlims,
  hard.min = -100000000000,
  hard.max = 100000000000,
  max.x.ticks = 6,
  abs.labels = FALSE
)
}
\arguments{
\item{xlims}{limits for X axis labels}

\item{hard.min}{hard limit for lower end of X-axis [default: -10e10]}

\item{hard.max}{hard limit for all components of X-axis [default: 10e10]}

\item{max.x.ticks}{maximum ticks for X axis [default: 6]}

\item{abs.labels}{print absolute value of axis labels [default: FALSE]}
}
\description{
Helper function to add an X-axis to a permutation test result plot
}
\seealso{
\code{\link[=plot.seg.perms]{plot.seg.perms()}}, \code{\link[=plot.seg.perms.multi]{plot.seg.perms.multi()}}, \code{\link[=plot.viohist]{plot.viohist()}}
}
