% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncoding_utils.R
\name{assign.track.family}
\alias{assign.track.family}
\title{Assign noncoding annotation families}
\usage{
assign.track.family(track, source)
}
\arguments{
\item{track}{track name}

\item{source}{name of track source}
}
\value{
string of track family
}
\description{
Assign noncoding annotation family based on annotation track & source
}
