% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{build.fisher.lookup.table}
\alias{build.fisher.lookup.table}
\title{Compute Fisher's exact test lookup table}
\usage{
build.fisher.lookup.table(bed)
}
\arguments{
\item{bed}{data frame of case/control CNV counts as loaded by \code{load.cc.cnv.counts()}}
}
\value{
data frame of all unique case & control CNV counts
}
\description{
Build a lookup table of all case/control count pairings for Fisher's Exact Tests
}
\seealso{
\code{\link[=load.cc.cnv.counts]{load.cc.cnv.counts()}}
}
