% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{calc.or}
\alias{calc.or}
\title{Calculate CNV odds ratio}
\usage{
calc.or(control_ref, control_alt, case_ref, case_alt, adj = 0.5)
}
\arguments{
\item{control_ref}{number of control non-carriers}

\item{control_alt}{number of control CNV carriers}

\item{case_ref}{number of case non-carriers}

\item{case_alt}{number of case CNV carriers}

\item{adj}{fixed value to use for continuity correction [default: 0.5]}
}
\value{
odds ratio (numeric)
}
\description{
Compute basic OR of CNV carriers with adjustment for zero-inflation
}
\details{
implements simple continuity correction with fixed value added to
all cells of contingency table
}
