% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{calc.perm.p}
\alias{calc.perm.p}
\title{Calculate & format permuted P-value}
\usage{
calc.perm.p(perm.vals, obs.val, alternative = "greater")
}
\arguments{
\item{perm.vals}{vector of results from precomputed permutations}

\item{obs.val}{empirically observed result}

\item{alternative}{specify sidedness of test (see \code{Details} below)}
}
\value{
list of two values:
\enumerate{
\item \verb{$p} : numerical P-value
\item \verb{$formatted} : P-value formatted for plotting with \code{rCNV2::format.pval()}
}
}
\description{
Computes a P-value by comparing an observed value to a vector of precomputed permutations
}
