% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{calc.segs.dat}
\alias{calc.segs.dat}
\title{Summarize a segment metric}
\usage{
calc.segs.dat(segs, feature, measure, subset_to_regions = NULL)
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{feature}{name of feature to evaluate}

\item{measure}{statistic to evaluate [default: mean]}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}
}
\value{
vector of statistic split by all, deletions, and duplications
}
\description{
Helper function to compute various statistics across a subset of segments
}
\details{
Valid options for \code{measure} are mean, median, sum, and frac.any
}
