% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_assoc_utils.R
\name{categorize.genes}
\alias{categorize.genes}
\title{Categorize fine-mapped genes}
\usage{
categorize.genes(credsets)
}
\arguments{
\item{credsets}{data.frame of credible sets as read by \code{\link[=load.credsets]{load.credsets()}}}
}
\value{
list of gene categories split by CNV type
}
\description{
Categorize genes based on their fine-mapped PIP and their rank in each credible set
}
\seealso{
\code{\link[=load.credsets]{load.credsets()}}
}
