% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{classify.genes.by.score}
\alias{classify.genes.by.score}
\title{Partition genes by dosage sensitivity}
\usage{
classify.genes.by.score(
  scores,
  hi.hc = 0.9,
  hi.lc = 0.5,
  ts.hc = 0.9,
  ts.lc = 0.5
)
}
\arguments{
\item{scores}{data.frame of gene scores as loaded by \code{\link[=load.scores]{load.scores()}}}

\item{hi.hc}{minimum score for high-confidence haploinsufficient genes [default: 0.9]}

\item{hi.lc}{minimum score for low-confidence haploinsufficient genes [default: 0.5]}

\item{ts.hc}{minimum score for high-confidence triplosensitive genes [default: 0.9]}

\item{ts.lc}{minimum score for low-confidence triplosensitive genes [default: 0.5]}
}
\value{
list of gene symbols for each group
}
\description{
Partition genes into subgroups based on dosage sensitivity scores
}
\seealso{
\code{\link[=load.scores]{load.scores()}}
}
