% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{combine.dnms}
\alias{combine.dnms}
\title{Combine dnm data from two cohorts}
\usage{
combine.dnms(d1, d2, mutrates)
}
\arguments{
\item{d1}{dnm data from first cohort as read by \code{\link[=load.dnms]{load.dnms()}}}

\item{d2}{dnm data from second cohort as read by \code{\link[=load.dnms]{load.dnms()}}}

\item{mutrates}{gene-specific mutation rates as read by \code{\link[=load.mutrates]{load.mutrates()}}}
}
\value{
data.frame of merged dnm data
}
\description{
Pool dnm data from two cohorts and update all residualized values
}
\seealso{
\code{\link[=load.dnms]{load.dnms()}}, \code{\link[=load.mutrates]{load.mutrates()}}
}
