% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{combine.single.cohort.assoc.stats}
\alias{combine.single.cohort.assoc.stats}
\title{Merge single-cohort association stats}
\usage{
combine.single.cohort.assoc.stats(stats.list)
}
\arguments{
\item{stats.list}{path to .tsv list of single-cohort association stats}
}
\value{
data frame of association stats for all cohorts
}
\description{
Merges a list of association statistics generated for multiple independent cohorts
}
\seealso{
\code{\link[=read.assoc.stats.single]{read.assoc.stats.single()}}
}
