% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_assoc_plot.R
\name{credsets.scatter}
\alias{credsets.scatter}
\title{Credible set scatterplot}
\usage{
credsets.scatter(
  credsets,
  x,
  y,
  subset_to_regions = NULL,
  xlims = NULL,
  ylims = NULL,
  add.lm = T,
  pt.cex = 1,
  horiz.lines.at = NULL,
  horiz.lines.lty = 1,
  horiz.lines.color = NULL,
  abline.a = NULL,
  abline.b = NULL,
  abline.lty = 1,
  blue.bg = TRUE,
  xtitle = NULL,
  x.title.line = 1.5,
  x.at = NULL,
  x.labs = NULL,
  x.labs.at = NULL,
  parse.x.labs = FALSE,
  x.title.cex = 1,
  ytitle = NULL,
  y.title.line = 1.7,
  y.at = NULL,
  y.labs = NULL,
  y.labs.at = NULL,
  parse.y.labs = FALSE,
  y.title.cex = 1,
  parmar = c(2.7, 2.7, 0.25, 0.25)
)
}
\arguments{
\item{credsets}{data.frame of credible sets as loaded by \code{\link[=load.credsets]{load.credsets()}}}

\item{x}{values to plot on X axis}

\item{y}{values to plot on Y axis}

\item{subset_to_regions}{vector of credset IDs to include [default: include all credsets]}

\item{xlims}{X axis limits}

\item{ylims}{Y axis limits}

\item{add.lm}{add linear trendline [default: TRUE]}

\item{pt.cex}{expansion factor for points [default: 1]}

\item{horiz.lines.at}{numeric vector indicating where to draw horizontal lines [default: no lines]}

\item{horiz.lines.lty}{lty parameter for \code{horiz.lines.at} [default: 1]}

\item{horiz.lines.color}{colors for \code{horiz.lines.at}}

\item{abline.a}{\code{a} value for embedded \code{\link[=abline]{abline()}} call}

\item{abline.b}{\code{b} value for embedded \code{\link[=abline]{abline()}} call}

\item{abline.lty}{lty parameter for embedded \code{\link[=abline]{abline()}} call}

\item{blue.bg}{add light blue background [default: TRUE]}

\item{xtitle}{title of X axis}

\item{x.title.line}{line of X axis title}

\item{x.at}{tick positions for X axis}

\item{x.labs}{labels for X axis ticks}

\item{x.labs.at}{if different from \code{x.at}, specify locations of \code{x.labs}}

\item{parse.x.labs}{format X axis labels with \code{\link[=parse]{parse()}} [default: FALSE]}

\item{x.title.cex}{expansion factor for X axis title [default: 1]}

\item{ytitle}{title of Y axis}

\item{y.title.line}{line of Y axis title}

\item{y.at}{tick positions for Y axis}

\item{y.labs}{labels for Y axis ticks}

\item{y.labs.at}{if different from \code{y.at}, specify locations of \code{y.labs}}

\item{parse.y.labs}{format Y axis labels with \code{\link[=parse]{parse()}} [default: FALSE]}

\item{y.title.cex}{expansion factor for Y axis title [default: 1]}

\item{parmar}{numeric vector of margins passed to \code{\link[=par]{par()}}}
}
\description{
Generic credsets scatterplot function
}
\seealso{
\code{\link[=load.credsets]{load.credsets()}}
}
