% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{dens.scatter}
\alias{dens.scatter}
\title{Density-colored scatterplot}
\usage{
dens.scatter(x, y, pt.cex = 1, parmar = rep(0.5, 4), add.cor = T)
}
\arguments{
\item{x}{independent variable vector}

\item{y}{dependent variable vector}

\item{pt.cex}{scaling factor for points}

\item{parmar}{argument for option(parmar=...)}

\item{add.cor}{boolean indicator to add correlation coefficient to plot [default: TRUE]}
}
\value{
None
}
\description{
Generate a 2D scatterplot colored by local density
}
\seealso{
\code{\link[=color.points.by.density]{color.points.by.density()}}
}
