% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{estimate.cohort.inflation}
\alias{estimate.cohort.inflation}
\title{Estimate cohort inflation terms}
\usage{
estimate.cohort.inflation(stats.df)
}
\arguments{
\item{stats.df}{association statistics for a single cohorts as read by \code{\link[=read.assoc.stats.single]{read.assoc.stats.single()}}}
}
\value{
named vector with mean standardized log odds ratio for each cohort
}
\description{
Estimates mean genome-wide standardized effect size per cohort
}
\details{
Note that \code{keep.or.confint} must be \code{TRUE} when reading data
with \code{\link[=read.assoc.stats.single]{read.assoc.stats.single()}} for this method to work
}
\seealso{
\code{\link[=combine.single.cohort.assoc.stats]{combine.single.cohort.assoc.stats()}}
}
