% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{evaluate.score}
\alias{evaluate.score}
\title{Evaluate performance of a single gene score}
\usage{
evaluate.score(stats, score, true.genes, false.genes)
}
\arguments{
\item{stats}{data.frame of scores for all genes}

\item{score}{score to evaluate; must be a column of \code{stats}}

\item{true.genes}{vector of gene symbols to consider true positives}

\item{false.genes}{vector of gene symbols to consider true negatives}
}
\value{
list of ROC and P-R statistics
}
\description{
Wrapper function to calculate all performance stats a single score
}
\seealso{
\code{\link[=roc]{roc()}}, \code{\link[=prc]{prc()}}
}
