% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{fisher.burden.test}
\alias{fisher.burden.test}
\title{Fisher's Exact Test for a list of loci}
\usage{
fisher.burden.test(bed, keep.n.cols = 3, precision = 10)
}
\arguments{
\item{bed}{data frame of case/control CNV counts as loaded by \code{load.cc.cnv.counts()}}

\item{keep.n.cols}{number of columns from original BED format to retain}

\item{precision}{maximum precision for floating point values}
}
\value{
data frame of association test resuts
}
\description{
Fisher's exact test of case:control CNV burden per locus for one or more loci
}
\seealso{
\code{\link[=fisher.burden.test.single]{fisher.burden.test.single()}}
}
