% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{fisher.burden.test.single}
\alias{fisher.burden.test.single}
\title{Single-locus Fisher's exact test}
\usage{
fisher.burden.test.single(counts, alternative = "greater")
}
\arguments{
\item{counts}{vector of CNV counts (see \code{Details})}

\item{alternative}{specify sidedness of test [default: 'greater']}
}
\value{
named vector with association summary statistics, in the order of
P-value, odds ratio, and lower/upper bounds of 95\% conf. int. of the OR
}
\description{
Fisher's exact test for a single vector of CNV counts for one cohort and one locus
}
\details{
\code{counts} must contain four values, in order:
\enumerate{
\item count of case CNV carriers
\item count of control CNV carriers
\item count of non-carrier cases
\item count of non-carrier controls
}
}
