% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{fit.exp.decay}
\alias{fit.exp.decay}
\title{Fit exponential decay}
\usage{
fit.exp.decay(x, y)
}
\arguments{
\item{x}{independent variable vector}

\item{y}{dependent variable vector}
}
\value{
Fitted model as \code{nls()} object with estimated parameters
}
\description{
Fits a univariate exponential decay with automatic parameter estimation
}
\details{
Based on example from \verb{https://rpubs.com/mengxu/exponential-model}
}
