% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{format.pval}
\alias{format.pval}
\title{Format P-value}
\usage{
\method{format}{pval}(p, nsmall = 2, max.decimal = 3, equality = "=", min.neg.log10.p = 100)
}
\arguments{
\item{p}{P-value}

\item{nsmall}{number of digits after the decimal to retain for scientific
notification [default: 2]}

\item{max.decimal}{convert all P-values requiring more digits after the decimal
to be converted to scientific notation [default: 3]}

\item{equality}{equality symbol to print after \code{P} [default: '=']}

\item{min.neg.log10.p}{minimum order of magnitude to process before considering
P-value to be arbitrarily/meaninglessly small [default: 100]}
}
\value{
formatted P-value as character
}
\description{
Format P-value for plotting
}
