% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{format.scientific}
\alias{format.scientific}
\title{Format scientific value}
\usage{
\method{format}{scientific}(x, nsmall = 2, max.decimal = 3)
}
\arguments{
\item{x}{numeric value to be formatted}

\item{nsmall}{number of digits after the decimal to retain for scientific
notification [default: 2]}

\item{max.decimal}{convert all values requiring more digits after the decimal
to be converted to scientific notation [default: 3]}
}
\value{
formatted value as character
}
\description{
Format a scientific value for plotting
}
