% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{gene.meta.otf}
\alias{gene.meta.otf}
\title{On-the-fly meta-analysis of a gene list}
\usage{
gene.meta.otf(meta.dat, genes, cnv)
}
\arguments{
\item{genes}{vector of genes of interest}

\item{cnv}{CNV type to evaluate (options: \code{DEL} or \code{DUP})}

\item{meta-dat}{meta-analysis inputs as loaded by \code{\link[=load.otf.meta.dat]{load.otf.meta.dat()}}}
}
\description{
Compute on-the-fly meta-analysis of all CNVs that hit a list of genes
}
\seealso{
\code{\link[=load.otf.meta.dat]{load.otf.meta.dat()}}
}
