% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{get.cytoband.range}
\alias{get.cytoband.range}
\title{Get cytoband range}
\usage{
get.cytoband.range(chrom, start, end, cytobands.in, drop.subbands = FALSE)
}
\arguments{
\item{chrom}{query chromosome}

\item{start}{left-most coordinate of query interval}

\item{end}{right-most coordinate of query interval}

\item{cytobands.in}{path to UCSC-style cytobands BED file}

\item{drop.subbands}{boolean indicator to round to the nearest whole band
[default: FALSE]}
}
\value{
character of cytoband range
}
\description{
Find & format the cytoband range overlapping a query interval
}
