% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{get.developmental.region_ids}
\alias{get.developmental.region_ids}
\title{Get list of developmental or adult loci}
\usage{
get.developmental.region_ids(loci, segs, adult = FALSE, sig.only = FALSE)
}
\arguments{
\item{loci}{locus association dataframe (imported with \code{\link[=load.loci]{load.loci()}})}

\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{adult}{return loci only associated with adult phenotypes [default: FALSE]}

\item{sig.only}{only return genome-wide or FDR significant segments [default: FALSE]}
}
\value{
vector
}
\description{
Compile master list of loci associated with developmental or adult phenotypes
}
\seealso{
\code{\link[=load.loci]{load.loci()}}, \code{\link[=load.segment.table]{load.segment.table()}}
}
