% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{get.gene.color.byscore}
\alias{get.gene.color.byscore}
\title{Color gene by dosage sensitivity}
\usage{
get.gene.color.byscore(gene, ds.groups)
}
\arguments{
\item{gene}{gene symbol to evaluate}

\item{ds.groups}{categories of high- and low-confidence genes as generated by \code{\link[=classify.genes.by.score]{classify.genes.by.score()}}}
}
\value{
color
}
\description{
Get gene color based on membership in a given dosage sensitivity subset
}
\seealso{
\code{\link[=classify.genes.by.score]{classify.genes.by.score()}}
}
