% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get.hpo.color}
\alias{get.hpo.color}
\title{Get HPO-based color code}
\usage{
get.hpo.color(hpo)
}
\arguments{
\item{hpo}{HPO code}
}
\value{
color code
}
\description{
Return hex color code for phenotype by HPO
}
