% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{get.neuro.region_ids}
\alias{get.neuro.region_ids}
\title{Get list of neuro-associated loci}
\usage{
get.neuro.region_ids(loci, segs)
}
\arguments{
\item{loci}{locus association dataframe (imported with \code{\link[=load.loci]{load.loci()}})}

\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}
}
\value{
vector
}
\description{
Compile master list of neuro-associated loci (using gw-sig for gw-sig loci and nominal for lit-based loci)
}
\seealso{
\code{\link[=load.loci]{load.loci()}}, \code{\link[=load.segment.table]{load.segment.table()}}
}
