% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_assoc_utils.R
\name{get.quadrant.genes}
\alias{get.quadrant.genes}
\title{Subset finemapped genes}
\usage{
get.quadrant.genes(gene.groups, top, conf)
}
\arguments{
\item{gene.groups}{categories of genes split by PIP & rank as produced by \code{\link[=categorize.genes]{categorize.genes()}}}

\item{top}{rank of genes in credible set}

\item{conf}{confidence level for subset}
}
\value{
vector of gene symbols
}
\description{
Get all genes for a single quadrant of 2x2 grid of fine-mapped genes
}
