% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{get.sample.counts}
\alias{get.sample.counts}
\title{Get sample counts from table}
\usage{
get.sample.counts(pheno.table.in, cohort.name, case.hpo, control.hpo)
}
\arguments{
\item{pheno.table.in}{path to phenotype table .tsv}

\item{cohort.name}{name of cohort}

\item{case.hpo}{HPO code for cases}

\item{control.hpo}{HPO code for controls}
}
\value{
list of two values:
\enumerate{
\item \verb{$case.n} : number of cases
\item \verb{$control.n} : number of controls
}
}
\description{
Extract case/control sample sizes from a single cohort for association testing
}
