% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{get.sample.sizes.for.highlight}
\alias{get.sample.sizes.for.highlight}
\title{Load sample sizes for locus highlights}
\usage{
get.sample.sizes.for.highlight(
  table.in,
  case.hpos = c("HP:0000118"),
  ctrl.hpo = "HEALTHY_CONTROL"
)
}
\arguments{
\item{table.in}{path to .tsv with sample sizes per HPO}

\item{case.hpos}{vector of case HPOs to evaluate [default: c("HP:0000118")]}

\item{ctrl.hpo}{control HPO [default: "HEALTHY_CONTROL"]}
}
\value{
list with the following two elements:
\enumerate{
\item \verb{$case} vector of case sample sizes
\item \verb{$ctrl} vector of control sample sizes
}
}
\description{
Load sample sizes for case/control contrast for locus highlight plotting
}
