% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_assoc_utils.R
\name{get.sig.gene.pw.params}
\alias{get.sig.gene.pw.params}
\title{Label gene associations by significance}
\usage{
get.sig.gene.pw.params(sig, cnv)
}
\arguments{
\item{sig}{character vector indicating significance level}

\item{cnv}{character vector of CNV type}
}
\value{
list of pt.bg and pt.border
}
\description{
Generate pointwise graphical annotations for gene associations by significance level
}
