% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{inv.var.avg}
\alias{inv.var.avg}
\title{Inverse-variance weighted mean}
\usage{
inv.var.avg(lnors, vars, conf = 0.95)
}
\arguments{
\item{vars}{variance for each entry in \code{values}}

\item{conf}{confidence interval to return [default: 0.95]}

\item{values}{numeric vector of values to be averaged}
}
\value{
vector of weighted mean and lower/upper bounds of confidence interval
}
\description{
Compute inverse-variance weighted mean and 95\% CI
}
