% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.bed3}
\alias{load.bed3}
\title{Load BED3}
\usage{
load.bed3(bed.in, header = T)
}
\arguments{
\item{bed.in}{path to input BED fil}

\item{header}{boolean indicator if BED has header [default: TRUE]}
}
\value{
data frame of coordinates
}
\description{
Load a simple BED3 file
}
