% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{load.cc.cnv.counts}
\alias{load.cc.cnv.counts}
\title{Load case/control CNV counts}
\usage{
load.cc.cnv.counts(bed.in, case.col.name, case.n, control.col.name, control.n)
}
\arguments{
\item{bed.in}{path to input BED3+ file}

\item{case.col.name}{name of column containing case CNV counts}

\item{case.n}{total number of case samples in cohort}

\item{control.col.name}{name of column containing control CNV counts}

\item{control.n}{total number of control samples in cohort}
}
\value{
data frame of CNV counts and frequencies per locus
}
\description{
Loads a file of CNV counts in cases and controls per locus
}
\details{
\code{bed.in} must adhere to format as generated by any of the following scripts:
\itemize{
\item \code{analysis/sliding_windows/count_cnvs_per_window.py}
\item \code{analysis/genes/count_cnvs_per_gene.py}
\item \code{analysis/noncoding/count_cnvs_per_crb.py}
}
}
