% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.chromhmm.colors}
\alias{load.chromhmm.colors}
\title{Load ChromHMM color code}
\usage{
load.chromhmm.colors(chromhmm.manifest.in)
}
\arguments{
\item{chromhmm.manifest.in}{path to ChromHMM manifest .tsv}
}
\value{
named vector of colors per ChromHMM state
}
\description{
Load ChromHMM state color code as HEX from manifest .tsv
}
