% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.chromhmm.tracks}
\alias{load.chromhmm.tracks}
\title{Load ChromHMM tracks for locus highlight}
\usage{
load.chromhmm.tracks(chromhmm.tracks.in, chromhmm.manifest.in, region)
}
\arguments{
\item{chromhmm.tracks.in}{.tsv of ChromHMM tracks to be loaded}

\item{chromhmm.manifest.in}{path to ChromHMM manifest .tsv}

\item{region}{coordinates of region to be extracted}
}
\value{
data.frame
}
\description{
Load a set of ChromHMM tracks from an input .tsv and apply color scheme
}
\seealso{
\link{load.chromhmm.colors}
}
