% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.cnvs.from.region.multi}
\alias{load.cnvs.from.region.multi}
\title{Load multiple CNV BEDs}
\usage{
load.cnvs.from.region.multi(
  cnvlist,
  region,
  cnv = NULL,
  case.hpo = "HP:0000118",
  ctrl.hpo = "HEALTHY_CONTROL",
  rstudio.local = FALSE
)
}
\arguments{
\item{cnvlist}{.tsv of CNV .bed files to load}

\item{region}{coordinates of region to be extracted}

\item{cnv}{filter to this CNV type [default: do not filter by CNV type]}

\item{ctrl.hpo}{control HPO [default: "HEALTHY_CONTROL"]}

\item{rstudio.local}{boolean to indicate local Rstudio environment [default: FALSE]}

\item{case.hpos}{vector of case HPOs to evaluate [default: c("HP:0000118")]}
}
\value{
list of outputs from \code{\link[=load.cnvs.from.region]{load.cnvs.from.region()}} for each row in \code{cnvlist}
}
\description{
Load all CNVs for a specific region from a list of BEDs
}
\details{
\code{region} coordinates must be a \code{tabix}-compatbile string
}
\seealso{
\link{load.cnvs.from.region}
}
