% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.cnvs.from.region}
\alias{load.cnvs.from.region}
\title{Load CNVs for a region of interest}
\usage{
load.cnvs.from.region(
  bedpath,
  region,
  cnv = NULL,
  case.hpos = c("HP:0000118"),
  ctrl.hpo = "HEALTHY_CONTROL",
  rstudio.local = FALSE
)
}
\arguments{
\item{bedpath}{path to .bed with CNVs}

\item{region}{coordinates of region to be extracted}

\item{cnv}{filter to this CNV type [default: do not filter by CNV type]}

\item{case.hpos}{vector of case HPOs to evaluate [default: c("HP:0000118")]}

\item{ctrl.hpo}{control HPO [default: "HEALTHY_CONTROL"]}

\item{rstudio.local}{boolean to indicate local Rstudio environment [default: FALSE]}
}
\value{
list with the following two elements:
\enumerate{
\item \verb{$case} data.frame of case CNVs
\item \verb{$ctrl} data.frame of control CNVs
}
}
\description{
Load CNVs from a single BED for a single region, and split by case/control
}
\details{
\code{region} coordinates must be a \code{tabix}-compatbile string
}
