% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncoding_utils.R
\name{load.crb.sumstats}
\alias{load.crb.sumstats}
\title{Load CRB sumstats}
\usage{
load.crb.sumstats(stats.in)
}
\arguments{
\item{stats.in}{path to .bed file of CRB association stats}
}
\description{
Extract summary statistics from a single gene or CRB BED
}
