% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_assoc_utils.R
\name{load.credsets}
\alias{load.credsets}
\title{Load credible sets}
\usage{
load.credsets(credsets.in)
}
\arguments{
\item{credsets.in}{path to BED file of credible sets}
}
\value{
data.frame
}
\description{
Load summary dataframe for credible set association stats
}
