% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.feature.bed.for.highlight}
\alias{load.feature.bed.for.highlight}
\title{Load quantitative feature}
\usage{
load.feature.bed.for.highlight(
  bedpath,
  region,
  keep.col = 4,
  rstudio.local = FALSE
)
}
\arguments{
\item{bedpath}{path to .bed file with feature information}

\item{region}{coordinates of region to be extracted}

\item{keep.col}{column number to use as feature values [default: 4]}

\item{rstudio.local}{boolean to indicate local Rstudio environment [default: FALSE]}
}
\value{
data.frame
}
\description{
Load a quantitative feature track from a BED file for locus highlights
}
\details{
\code{region} coordinates must be a \code{tabix}-compatbile string
}
