% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.features}
\alias{load.features}
\title{Load gene features BED}
\usage{
load.features(features.in, fill = NA, norm = F)
}
\arguments{
\item{features.in}{path to BED4+ file with feature columns (see \code{Details})}

\item{fill}{specify fill behavior for missing values. See \code{Details} for
options. [default: NA]}

\item{norm}{boolean indicator whether standard normalization should be
applied to features upon loading. [default: FALSE]}
}
\value{
data frame of features
}
\description{
Load a BED4+ formatted .tsv of gene features
}
\details{
\code{features.in} must be BED4+, with column 4 corresponding to gene symbol

\code{fill} accepts the following values:
\itemize{
\item \code{mean} : fill missing values with column-wise mean
\item \code{median} : fill missing values with column-wise median
\item \code{{numeric}} : fill missing values with a user-specified numeric constant
\item \code{NA} : do not fill missing values
}
}
