% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_assoc_utils.R
\name{load.gene.associations}
\alias{load.gene.associations}
\title{Load gene associations}
\usage{
load.gene.associations(assocs.in)
}
\arguments{
\item{assocs.in}{path to BED file of gene-level association stats}
}
\value{
data.frame
}
\description{
Load all gene association stats from a BED file
}
