% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R, R/gene_score_utils.R
\name{load.gene.feature.metadata}
\alias{load.gene.feature.metadata}
\title{Load gene feature metadata}
\usage{
load.gene.feature.metadata(meta.in)

load.gene.feature.metadata(meta.in)
}
\arguments{
\item{meta.in}{path to metadata .tsv}

\item{feature.metadata.in}{path to gene feature metadata .tsv}
}
\value{
data.frame of contents from \code{feature.metadata.in}

data.frame
}
\description{
Load table of gene feature metadata (including plain-text descriptions)

Load gene feature category metadata from .tsv
}
