% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.gene.lists}
\alias{load.gene.lists}
\title{Load a list of gene lists}
\usage{
load.gene.lists(genelists.in)
}
\arguments{
\item{genelists.in}{input .tsv (see \code{details})}
}
\value{
list of character vectors, one per row in \code{genelists.in}
}
\description{
Loads a list of gene lists from an input .tsv
}
\details{
\code{genelists.in} must be a .tsv with the following two columns:
\enumerate{
\item gene list name
\item path to flat .txt file with all genes in list
}
}
