% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{load.gene.metadata}
\alias{load.gene.metadata}
\title{Load gene metadata}
\usage{
load.gene.metadata(cov.in)
}
\arguments{
\item{cov.in}{path to gene metadata}
}
\value{
data.frame
}
\description{
Load gene metadata from a BED file
}
