% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.genelist}
\alias{load.genelist}
\title{Load gene list}
\usage{
load.genelist(path)
}
\arguments{
\item{path}{path to gene list}
}
\value{
character vector of gene symbols
}
\description{
Loads a single list of genes from a flat text file
}
