% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.genes.from.gtf}
\alias{load.genes.from.gtf}
\title{Load genes from GTF for plotting}
\usage{
load.genes.from.gtf(gtf.in, region, rstudio.local = FALSE)
}
\arguments{
\item{gtf.in}{path to .gtf}

\item{region}{coordinates of region to be extracted}

\item{rstudio.local}{boolean to indicate local Rstudio environment [default: FALSE]}
}
\value{
data.frame
}
\description{
Extract gene features from a GTF for plotting
}
\details{
\code{region} coordinates must be a \code{tabix}-compatbile string
}
