% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.hpo.jaccard.matrix}
\alias{load.hpo.jaccard.matrix}
\title{Load HPO Jaccard matrix}
\usage{
load.hpo.jaccard.matrix(hpo.jac.in)
}
\arguments{
\item{hpo.jac.in}{Tab-delimited matrix of Jaccard indexes for all HPO pairs}
}
\value{
data.frame
}
\description{
Load a square matrix of Jaccard indexes for all HPO pairs
}
