% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_plot_utils.R
\name{load.hpo.metadata}
\alias{load.hpo.metadata}
\title{Load HPO metadata}
\usage{
load.hpo.metadata(meta.in, counts.in, hpos)
}
\arguments{
\item{meta.in}{path to metadata .tsv}

\item{counts.in}{path to .tsv of sample counts}

\item{hpos}{character vector specifying the desired order of output}
}
\value{
data frame of metadata
}
\description{
Load a table of HPO metadata
}
