% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{load.loci}
\alias{load.loci}
\title{Load locus-level association stats}
\usage{
load.loci(loci.in)
}
\arguments{
\item{loci.in}{path to loci BED file}
}
\value{
data.frame
}
\description{
Load summary dataframe for locus-level association stats
}
