% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_qq.R
\name{load.manhattan.stats}
\alias{load.manhattan.stats}
\title{Load association statistics}
\usage{
load.manhattan.stats(
  stats.in,
  p.col.name = "p",
  p.is.phred = F,
  min.p = 10^-100
)
}
\arguments{
\item{stats.in}{path to input file of association statistics}

\item{p.col.name}{name of column in header for P-value column [default: 'p']}

\item{p.is.phred}{boolean indicator if P-values as provided are -log10 scaled [default: FALSE]}

\item{min.p}{minimum P-value to retain as numerically meaningful [default: 10^-100]}
}
\value{
data frame with three columns: chromosome, position, and P-value
}
\description{
Loads a generic input file of locus-based association statistics
}
