% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.meta.stats}
\alias{load.meta.stats}
\title{Load meta-analysis association statistics}
\usage{
load.meta.stats(stats.in, p.is.neg.log10, keep.n.cols = 3)
}
\arguments{
\item{stats.in}{path to input BED file with association statistics}

\item{keep.n.cols}{number of columns from original BED format to retain}
}
\value{
data frame of formatted association stats
}
\description{
Load association statistics from a meta-analysis
}
