% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.mutrates}
\alias{load.mutrates}
\title{Load mutation rates}
\usage{
load.mutrates(mutrates.in)
}
\arguments{
\item{mutrates.in}{.tsv of mutation rates per gene}
}
\value{
data.frame of mutation rates
}
\description{
Load a table of gene-specific mutation rates
}
