% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{load.otf.cohort.cnvs}
\alias{load.otf.cohort.cnvs}
\title{Load single cohort data for on-the-fly meta-analysis}
\usage{
load.otf.cohort.cnvs(cnvs.in)
}
\arguments{
\item{cnvs.in}{path to input .tsv (see \code{details})}
}
\value{
data.frame with two columns:
\itemize{
\item \verb{$iscase} boolean for whether CNV was observed in case sample
\item \verb{$genes} list of genes impacted by this CNV
}
}
\description{
Load a list of genes impacted per CNV for a single cohort for on-the-fly meta-analyses
}
\details{
\code{cnvs.in} must be a .tsv with at least the following two columns:
\enumerate{
\item \code{phenos} : semicolon-delimited list of all HPO terms corresponding to that CNV
\item \code{genes} : semicolon-delimited list of genes impacted by that CNV
Other columns can be optionally present and will be ignored.
}
}
