% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{load.other.scores}
\alias{load.other.scores}
\title{Load non-rCNV gene scores}
\usage{
load.other.scores(meta.in)
}
\arguments{
\item{meta.in}{path to gene constraint metadata .tsv}
}
\value{
data.frame of scores
}
\description{
Load & normalize other (non-rCNV) scores
}
