% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{load.perms.bygene}
\alias{load.perms.bygene}
\title{Load gene-based permutation data}
\usage{
load.perms.bygene(perm.res.in, subset_to_regions = NULL)
}
\arguments{
\item{perm.res.in}{path to .tsv of all permutation results}

\item{subset_to_regions}{optional vector of region IDs to keep [default: keep all regions]}
}
\value{
data.table() of permutation results
}
\description{
Load all permuted segments matched on the number of genes
}
