% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.pips.for.genelist}
\alias{load.pips.for.genelist}
\title{Load PIPs for a list of genes}
\usage{
load.pips.for.genelist(pips.in, genes, hpos = NULL)
}
\arguments{
\item{pips.in}{path to .tsv of PIPs per gene}

\item{genes}{vector of gene symbols to retain}

\item{hpos}{HPO(s) of interest [default: do not filter on HPO]}
}
\value{
data.frame
}
\description{
Extract PIPs from BED for specified genes for locus highlight plotting
}
