% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{load.score.cutoffs}
\alias{load.score.cutoffs}
\title{Load gene score cutoffs}
\usage{
load.score.cutoffs(
  score.cutoffs.in,
  default.hc.cutoff = 0.9,
  default.lc.cutoff = 0.5
)
}
\arguments{
\item{score.cutoffs.in}{path to input .tsv with score cutoffs}

\item{default.hc.cutoff}{default high-confidence cutoff [default: 0.9]}

\item{default.lc.cutoff}{default low-confidence cutoff [default: 0.5]}
}
\description{
Load gene score cutoffs from a .tsv into global environment
}
\details{
expects a .tsv as generated by empirical_score_cutoffs.R.
If \code{score.cutoffs.in} is \code{NULL}, will load default cutoffs.
}
