% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{load.scores}
\alias{load.scores}
\title{Load gene scores}
\usage{
load.scores(scores.in)
}
\arguments{
\item{scores.in}{path to gene scores .tsv}
}
\value{
data.frame
}
\description{
Load a dataframe of all gene scores
}
