% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{load.segment.table}
\alias{load.segment.table}
\title{Load main segment table}
\usage{
load.segment.table(segs.in)
}
\arguments{
\item{segs.in}{path to segments BED file}
}
\value{
data.frame
}
\description{
Load summary table for all segments (including lit. GDs & NAHR regions)
}
