% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncoding_utils.R
\name{load.track.stats}
\alias{load.track.stats}
\title{Load track burden sumstats}
\usage{
load.track.stats(stats.in, spa = F)
}
\arguments{
\item{stats.in}{path to track burden summary statistics file}

\item{spa}{update P-values with saddlepoint re-approximation of the null [default: FALSE]}
}
\value{
data.frame
}
\description{
Load burden test statistics for all tracks
}
