% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{make.meta.df}
\alias{make.meta.df}
\title{Create meta-analysis data frame}
\usage{
make.meta.df(stats.merged, cohorts, row.idx, empirical.continuity = T)
}
\arguments{
\item{stats.merged}{association statistics for all cohorts (see \code{\link[=combine.single.cohort.assoc.stats]{combine.single.cohort.assoc.stats()}})}

\item{cohorts}{vector of all cohort names}

\item{row.idx}{row index in \code{stats.merged} corresponding to locus of interest}

\item{empirical.continuity}{boolean indicator to apply Sweeting empirical
continuity correction [default: TRUE]}
}
\value{
data frame of statistics ready for meta-analysis
}
\description{
Make meta-analysis data frame for CNV association testing of a single locus
}
\seealso{
\code{\link[=combine.single.cohort.assoc.stats]{combine.single.cohort.assoc.stats()}}
}
