% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{merge.loci.segs}
\alias{merge.loci.segs}
\title{Merge locus and segment files}
\usage{
\method{merge}{loci.segs}(loci, segs)
}
\arguments{
\item{loci}{locus association dataframe (imported with \code{\link[=load.loci]{load.loci()}})}

\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}
}
\value{
vector
}
\description{
Merge locus association data with master segments table
}
\seealso{
\code{\link[=load.loci]{load.loci()}}, \code{\link[=load.segment.table]{load.segment.table()}}
}
