% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{meta.single}
\alias{meta.single}
\title{Single locus meta-analysis}
\usage{
meta.single(
  stats.merged,
  cohorts,
  row.idx,
  model = "fe",
  empirical.continuity = T,
  drop_top_cohort = F
)
}
\arguments{
\item{stats.merged}{association statistics for all cohorts (see \code{\link[=combine.single.cohort.assoc.stats]{combine.single.cohort.assoc.stats()}})}

\item{cohorts}{vector of all cohort names}

\item{row.idx}{row index in \code{stats.merged} corresponding to locus of interest}

\item{model}{specify meta-analysis model to use (see \code{Details})}

\item{empirical.continuity}{boolean indicator to apply Sweeting empirical
continuity correction [default: TRUE]}

\item{drop_top_cohort}{boolean indicator to drop most significant individual
cohort from meta-analysis [default: FALSE]}
}
\value{
numeric vector of meta-analysis summary statistics
}
\description{
Perform a meta-analysis of CNVs for a single locus
}
\details{
\code{model} accepts several character inputs, including:
\itemize{
\item \code{fe} : fixed-effects model implemented by \code{\link[metafor:rma.uni]{metafor::rma.uni()}} [default]
\item \code{re} : random-effects model implemented by \code{\link[metafor:rma.uni]{metafor::rma.uni()}}
\item \code{mh} : Mantel-Haenszel model implemented by \code{\link[metafor:rma.mh]{metafor::rma.mh()}}
}
}
\seealso{
\code{\link[=make.meta.df]{make.meta.df()}}
}
