% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{mod.features}
\alias{mod.features}
\title{Make gene features more interpretable}
\usage{
mod.features(feats)
}
\arguments{
\item{feats}{data.frame of gene features}
}
\value{
data.frame of modified gene features
}
\description{
Modify gene feature matrix to improve interpretability of regression coefficients
}
