% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{normalize.dnms}
\alias{normalize.dnms}
\title{Normalize DNM counts}
\usage{
normalize.dnms(segs, dnm.cohorts = c("DDD", "ASC", "ASC_unaffected"))
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{dnm.cohorts}{vector of cohorts to evaluate (default: DDD, ASC, ASC_unaffected)}
}
\value{
data.frame
}
\description{
Normalize segment DNM counts vs. synonymous inflation
}
