% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{or.corplot.grid}
\alias{or.corplot.grid}
\title{Multi-cohort pairwise effect size comparison}
\usage{
or.corplot.grid(stats.list, pt.cex = 1)
}
\arguments{
\item{stats.list}{path to .tsv list of single-cohort association stats}

\item{pt.cex}{scaling factor for points}
}
\value{
None
}
\description{
Make a grid of scatterplots of log odds ratios pairwise for all cohorts in a meta-analysis
}
\seealso{
\code{\link[=read.assoc.stats.single]{read.assoc.stats.single()}}
}
