% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{parse.region.for.highlight}
\alias{parse.region.for.highlight}
\title{Parse region for locus highlight}
\usage{
parse.region.for.highlight(
  region,
  genome.in,
  export.values = TRUE,
  return.values = FALSE
)
}
\arguments{
\item{region}{tabix-style string of region to be plotted}

\item{genome.in}{path to BEDTools-style .genome file}

\item{export.values}{boolean indicator to export values to .GlobalEnv [default: TRUE]}

\item{return.values}{boolean indicator to return values as vector [default: FALSE]}
}
\value{
vector (if \code{return.values} is \code{TRUE})
}
\description{
Parse a tabix-style coordinate string for locus highlight
}
\details{
if necessary, start and end coordinates will be bounded to the global
start and end coordinates of that chromosome as per \code{genome.in}
}
