% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{plot.all.perm.res}
\alias{plot.all.perm.res}
\title{Plot all segment permutation results}
\usage{
\method{plot}{all.perm.res}(
  segs,
  gw.perms,
  lit.perms,
  feature,
  measure,
  outdir,
  prefix,
  invert = F,
  subset_to_regions = NULL,
  norm = F,
  norm.multi = F,
  return.stats = TRUE,
  print.stats = FALSE,
  n.bins.single = 100,
  n.bins.multi = 100,
  min.bins = 10,
  x.title = NULL,
  xlims = NULL,
  xmin = NULL,
  xmax = NULL,
  pdf.dims.single = c(2.2, 2.4),
  parmar.single = c(2.25, 2, 0.5, 0.5),
  pdf.dims.multi = c(2.2, 4.8),
  parmar.multi = c(2.25, 6, 0.5, 0.5)
)
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{gw.perms}{list of permutation results for empirical segments}

\item{lit.perms}{list of permutation results for literature segments}

\item{feature}{name of feature to evaluate}

\item{measure}{statistic to evaluate}

\item{outdir}{path to output directory}

\item{prefix}{filename prefix for output plots}

\item{invert}{invert orientation of X-axis [default: FALSE]}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}

\item{norm}{normalize values before plotting [default: FALSE]}

\item{norm.multi}{normalize values before plotting for multi-panel figure [default: FALSE]}

\item{return.stats}{return dataframe of summary stats [default: TRUE]}

\item{print.stats}{print summary of stats [default: FALSE]}

\item{n.bins.single}{ideal number of bins for single-panel histograms [default: 100]}

\item{n.bins.multi}{ideal number of bins for multi-panel histograms [default: 100]}

\item{min.bins}{minimum number of bins [default: 10]}

\item{xlims}{limits for X axis}

\item{xmin}{minimum for X axis}

\item{xmax}{maximum for X axis}

\item{pdf.dims.single}{dimensionf of pdf for single-panel plots}

\item{parmar.single}{numeric vector of margins passed to \code{\link[=par]{par()}} for single-panel plots}

\item{pdf.dims.multi}{dimensions for multi-panel plots}

\item{parmar.multi}{numeric vector of margins passed to \code{\link[=par]{par()}} for multi-panel plots}

\item{xtitle}{title of X axis}

\item{x.title.line}{line of X axis title}
}
\description{
Wrapper function to plot four views on segment permutation results
}
\seealso{
\code{\link[=plot.seg.perms.multi]{plot.seg.perms.multi()}}
}
