% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.bracket}
\alias{plot.bracket}
\title{Plot square bracket}
\usage{
\method{plot}{bracket}(
  xleft,
  xright,
  y0,
  height,
  col = blueblack,
  staple.wex = 0.025,
  lwd = 1,
  left.label = NULL
)
}
\arguments{
\item{xleft}{left-most coordinate for bracket}

\item{xright}{right-most coordinate for bracket}

\item{y0}{vertical midpoint for bracket}

\item{height}{height of bracket}

\item{col}{color for bracket [default: blueblack]}

\item{staple.wex}{relative width of bracket staples [default: 0.025]}

\item{lwd}{line width [default: 1]}

\item{left.label}{label to be printed at the left end of the bracket
[default: no label]}
}
\description{
Add square bracket to locus highlight
}
