% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.chromhmm.tracks}
\alias{plot.chromhmm.tracks}
\title{Plot ChromHMM tracks}
\usage{
\method{plot}{chromhmm.tracks}(chmm.tracks, y0, panel.height = 0.2, y.axis.title = NULL)
}
\arguments{
\item{chmm.tracks}{list of chromhmm tracks to be plotted as loaded by []}

\item{y0}{vertical midpoint for panel}

\item{panel.height}{relative height of panel [default: 0.2]}

\item{y.axis.title}{title for Y axis [default: no title]}
}
\description{
Add panel of ChromHMM tracks to locus highlight
}
\seealso{
\link{load.chromhmm.tracks}
}
