% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.cnv.key.for.highlight}
\alias{plot.cnv.key.for.highlight}
\title{Plot CNV key for locus highlight}
\usage{
\method{plot}{cnv.key.for.highlight}(
  cnv.type,
  y0,
  total.n.ctrls,
  all.case.hpos,
  total.n.cases,
  highlight.case.hpo = NA,
  total.n.cases.highlight = NA,
  panel.height = 0.2,
  text.cex = 5/6,
  pt.cex = 1.3
)
}
\arguments{
\item{cnv.type}{type of CNV}

\item{y0}{midpoint for key panel}

\item{total.n.ctrls}{total number of all controls}

\item{all.case.hpos}{all HPOs to be considered as cases}

\item{total.n.cases}{total number of all cases}

\item{highlight.case.hpo}{HPO to be highlighted [default: no highlight]}

\item{total.n.cases.highlight}{total number of cases matching \code{highlight.case.hpo}
[default no highlight]}

\item{panel.height}{relative height of key panel [default: 0.2]}

\item{text.cex}{character expansion scalar for text [default: 5/6]}

\item{pt.cex}{expansion scalar for points [default: 1.3]}
}
\description{
Add CNV key to locus highlight
}
