% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.cnv.panel.for.highlight}
\alias{plot.cnv.panel.for.highlight}
\title{Plot CNV panel for locus highlight}
\usage{
\method{plot}{cnv.panel.for.highlight}(
  cnvs,
  n.case,
  n.ctrl,
  y0,
  cnv.type,
  highlight.hpo = NA,
  max.freq = NULL,
  start = NULL,
  end = NULL,
  y.axis.title = "CNV\\nFreq.",
  expand.pheno.label = TRUE,
  case.legend.side = "left",
  case.legend.topbottom = "top",
  ctrl.legend.side = "left",
  cc.legend.colors = rep(blueblack, 2),
  cnv.lwd = 0,
  add.cohort.label = FALSE,
  cohort.label = NULL,
  panel.height = 2,
  dx = 100
)
}
\arguments{
\item{cnvs}{cnvs to plot as loaded by \code{\link[=load.cnvs.from.region]{load.cnvs.from.region()}}}

\item{n.case}{case sample size}

\item{n.ctrl}{control sample size}

\item{y0}{vertical midpoint for panel}

\item{cnv.type}{type of CNV to plot}

\item{highlight.hpo}{highlight CNVs from this HPO in a different color
[default: plot all CNVs in the same color]}

\item{max.freq}{maximum frequency to display on Y-axis [default: auto-set]}

\item{start}{left-most coordinate to plot [default: plot all CNVs]}

\item{end}{right-most coordinate to plot [default: plot all CNVs]}

\item{y.axis.title}{title for Y axis [default: "CNV\\nFreq."]}

\item{expand.pheno.label}{boolean indicator whether to expand phenotype label [default: TRUE]}

\item{case.legend.side}{side to plot case legend [default: "left"]}

\item{case.legend.topbottom}{vertical alignment for case legend [default "top"]}

\item{ctrl.legend.side}{side to plot control legend [default: "left"]}

\item{cc.legend.colors}{vector of colors to apply to case & control legends [default: blueblack]}

\item{cnv.lwd}{line width for each CNV [default: 0]}

\item{add.cohort.label}{boolean indicator to print \code{cohort.label} on plot [default: FALSE]}

\item{cohort.label}{text of cohort label [default: no label]}

\item{panel.height}{relative height of panel [default: 2]}

\item{dx}{number of steps across X axis [default: 100]}
}
\description{
Add mirrored case & control CNV pileups for a single cohort to a locus highlight
}
\details{
\code{case.legend.side} and \code{control.legend.side} both accept either
"top" or "bottom".
Vertical alignment of control legend will automatically be set to opposite
of \code{case.legend.topbottom}.
}
\seealso{
\link{load.cnvs.from.region}, \link{pileup.cnvs.for.highlight}
}
