% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.constraint.track}
\alias{plot.constraint.track}
\title{Plot constraint track for locus highlight}
\usage{
\method{plot}{constraint.track}(constr.df, genes, y0, panel.height = 0.2, label.genes = NULL)
}
\arguments{
\item{constr.df}{data.frame of constraint features per gene as loaded by \code{\link[=load.features]{load.features()}}}

\item{genes}{data.frame of gene features to plot as loaded by \code{\link[=load.genes.from.gtf]{load.genes.from.gtf()}}}

\item{y0}{vertical midpoint for panel}

\item{panel.height}{relative height of panel [default: 0.2]}

\item{label.genes}{vector of gene symbols to label [default: don't label genes]}
}
\description{
Add plot of LOEUF values for all genes for locus highlight
}
\seealso{
\link{load.features}, \link{load.genes.from.gtf}
}
