% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.coord.line}
\alias{plot.coord.line}
\title{Plot coordinate line}
\usage{
\method{plot}{coord.line}(
  start,
  end,
  y0,
  highlight.start,
  highlight.end,
  highlight.col = "red",
  tick.height = 0.1,
  max.tick = 6,
  vlines = FALSE,
  vlines.bottom = NULL,
  lab.cex = 4.5/6
)
}
\arguments{
\item{start}{left-most coordinate}

\item{end}{right-most coordinate}

\item{y0}{position on Y axis for coordinate line}

\item{highlight.start}{left-most coordinate(s) to highlight}

\item{highlight.end}{right-most coordinate(s) to highlight}

\item{highlight.col}{color for highlights [default: "red"]}

\item{tick.height}{relative height for vertical ticks [default: 0.1]}

\item{max.tick}{maximum number of ticks across the entire coordinate line [default 6]}

\item{vlines}{boolean indicator to add vertical lines for each tick [default: FALSE]}

\item{vlines.bottom}{lowest Y coordinate to extend \code{vlines} [default: bottom of user coordinates]}

\item{lab.cex}{character expansion scalar for labels [default: 4.5/6]}
}
\description{
Add a rounded coordinate line to a locus highlight plot
}
