% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_plot_utils.R
\name{plot.dendro}
\alias{plot.dendro}
\title{Plot HPO dendrogram}
\usage{
\method{plot}{dendro}(meta, lwd = 1, color = "black", y.top = -0.6, box.wex = 1)
}
\arguments{
\item{meta}{HPO metadata \(see \code{\link[=load.hpo.metadata]{load.hpo.metadata()}}\)}

\item{lwd}{line width}

\item{color}{line color}

\item{y.top}{top Y coordinate}
}
\description{
Plot a vertical dendrogram of HPO hierarchy
}
