% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_plot.R
\name{plot.feature.bydsgroup}
\alias{plot.feature.bydsgroup}
\title{Plot gene feature distribution vs. scores}
\usage{
\method{plot}{feature.bydsgroup}(
  feats,
  ds.groups,
  hc.lab = "High Conf.",
  lc.lab = "Low Conf.",
  feat.idx = 2,
  title = NULL,
  swarm.max = 5000,
  max.ylim = 6,
  blue.bg = FALSE,
  parmar = c(2.25, 2.75, 1.25, 0.5)
)
}
\arguments{
\item{feats}{matrix of gene features}

\item{ds.groups}{list of genes split by score category as generated by \code{\link[=classify.genes.by.score]{classify.genes.by.score()}}}

\item{hc.lab}{label for high confidence genes [default: "High Conf."]}

\item{lc.lab}{label for low confidence genes [default: "Low Conf."]}

\item{feat.idx}{column index of feature to be plotted [default: 2]}

\item{title}{title for plot [default: NULL]}

\item{swarm.max}{maximum number of points for a swarmplot before rendering as a violin [default: 5000]}

\item{max.ylim}{maximum Y axis value [default: 6]}

\item{blue.bg}{add light blue background shading [default: FALSE]}

\item{parmar}{plotting margins passed to \code{par(mar=...)}}
}
\description{
Swarmplots of distributions for a single gene feature per gene score category
}
\seealso{
\code{\link[=load.scores]{load.scores()}}, \code{\link[=classify.genes.by.score]{classify.genes.by.score()}}
}
