% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.features.from.bedgraph}
\alias{plot.features.from.bedgraph}
\title{Plot quantitative bedgraph-style feature}
\usage{
\method{plot}{features.from.bedgraph}(bed, y0, col = blueblack, panel.height = 0.2, ytitle = NULL)
}
\arguments{
\item{bed}{BED-style data.frame of elements to plot as loaded by
\code{\link[=load.feature.bed.for.highlight]{load.feature.bed.for.highlight()}}}

\item{y0}{vertical midpoint for panel}

\item{col}{color to fill bars}

\item{panel.height}{relative height of panel [default: 0.2]}

\item{y.axis.title}{title for Y axis [default: no title]}
}
\description{
Add track of BEDGraph-style quantitative feature as coordinate-based barplot to locus highlight
}
\seealso{
\link{load.feature.bed.for.highlight}
}
