% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.gene.bodies}
\alias{plot.gene.bodies}
\title{Plot gene bodies}
\usage{
\method{plot}{gene.bodies}(
  genes,
  y0,
  transcripts = FALSE,
  mark.tss = FALSE,
  n.rows = 1,
  panel.height = 0.2,
  col = ns.color,
  y.axis.title = NULL,
  y.axis.title.col = NULL,
  label.genes = c()
)
}
\arguments{
\item{genes}{gene features to be plotted, loaded by \code{\link[=load.genes.from.gtf]{load.genes.from.gtf()}}}

\item{y0}{center position on Y axis for gene panel}

\item{transcripts}{boolean indicator to plot all transcripts per gene
[default: collapse all transcripts per gene]}

\item{mark.tss}{boolean indicator to mark TSS of each gene [default: FALSE]}

\item{n.rows}{number of rows of genes [default: 1]}

\item{panel.height}{relative height of gene panel [default: 0.2]}

\item{col}{color for genes [default: ns.color]}

\item{y.axis.title}{title for Y axis [default: no title]}

\item{y.axis.title.col}{color for \code{y.axis.title} [default: no title]}

\item{label.genes}{vector of gene symbols to be labeled on plot}
}
\description{
Add gene bodies to locus highlight
}
\seealso{
\link{load.genes.from.gtf}
}
