% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.na.rect}
\alias{plot.na.rect}
\title{Plot hatched N/A rectangle}
\usage{
\method{plot}{na.rect}(xleft, xright, y0, panel.height = 0.2, text = NULL)
}
\arguments{
\item{xleft}{left-most coordinate(s) for rectangles}

\item{xright}{right-most coordinate(s) for rectangles}

\item{y0}{vertical midpoint for panel}

\item{panel.height}{relative height of panel [default: 0.2]}

\item{text}{text label to superimpose on top of rectangles [default: no text]}
}
\description{
Add hatched NA rectangle to locus highlight
}
