% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.ors.for.highlight}
\alias{plot.ors.for.highlight}
\title{Plot odds ratios for locus highlight}
\usage{
\method{plot}{ors.for.highlight}(ss, y0, cnv.type, panel.height = 0.2, pt.cex = 0.7)
}
\arguments{
\item{ss}{summary statistics to be plotted as loaded by \link{load.sumstats.for.region}}

\item{y0}{vertical midpoint for panel}

\item{cnv.type}{CNV type to be plotted}

\item{panel.height}{relative height of panel [default: 0.2]}

\item{pt.cex}{character expansion scalar for points [default: 0.7]}
}
\description{
Add panel of log-odds ratios to locus highlight
}
\seealso{
\link{load.sumstats.for.region}
}
