% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.pips.for.highlight}
\alias{plot.pips.for.highlight}
\title{Plot PIPs for locus highlight}
\usage{
\method{plot}{pips.for.highlight}(
  pips,
  genes,
  y0,
  panel.height = 0.2,
  label.genes = NULL,
  col = blueblack,
  highlight.genes = NULL,
  highlight.col = "red"
)
}
\arguments{
\item{pips}{data.frame of PIPs per gene as loaded by \code{\link[=load.pips.for.genelist]{load.pips.for.genelist()}}}

\item{genes}{data.frame of gene features to plot as loaded by \code{\link[=load.genes.from.gtf]{load.genes.from.gtf()}}}

\item{y0}{vertical midpoint for panel}

\item{panel.height}{relative height of panel [default: 0.2]}

\item{label.genes}{vector of gene symbols to label [default: don't label genes]}

\item{col}{color for plotting genes [default: blueblack]}

\item{highlight.genes}{vector of gene symbols to plot in a different color
[default: plot all genes using \code{col}]}

\item{highlight.col}{color to apply to \code{highlight.genes} [default: "red"]}
}
\description{
Add plot of PIPs for all genes for locus highlight
}
\seealso{
\link{load.pips.for.genelist}, \link{load.genes.from.gtf}
}
