% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.pvalues.for.highlight}
\alias{plot.pvalues.for.highlight}
\title{Plot P-values for locus highlight}
\usage{
\method{plot}{pvalues.for.highlight}(
  ss,
  y0,
  cnv.type,
  panel.height = 0.2,
  pt.cex = 0.6,
  gw.sig = NULL,
  min.y = 9,
  gw.sig.label = "Genome-wide significance",
  gw.sig.label.side = "above"
)
}
\arguments{
\item{ss}{summary statistics to be plotted as loaded by \link{load.sumstats.for.region}}

\item{y0}{vertical midpoint for panel}

\item{cnv.type}{CNV type to be plotted}

\item{panel.height}{relative height of panel [default: 0.6]}

\item{pt.cex}{character expansion scalar for points [default: 0.6]}

\item{gw.sig}{value to annotate as genome-wide significance
[default: do not annotate genome-wide significance]}

\item{min.y}{minimum -log10(P) for the maximum Y axis [default: 9]}

\item{gw.sig.label}{text label for \code{gw.sig} [default "Genome-wide significance"]}

\item{gw.sig.label.side}{orientation of \code{gw.sig.label} relative to \code{gw.sig} line
[default: "above"]}
}
\description{
Add panel of neglog10-scaled P-values to locus highlight
}
\details{
\code{gw.sig.label.side} accepts either "above" or "below"
}
\seealso{
\link{load.sumstats.for.region}
}
