% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_qq.R
\name{plot.qq}
\alias{plot.qq}
\title{Quantile-quantile plot}
\usage{
\method{plot}{qq}(
  stats,
  cutoff = NULL,
  highlights = NULL,
  highlight.color = "#4EE69A",
  highlight.name = "Positive Controls",
  print.stats = T,
  echo.lambdas = F,
  legend = T,
  xmax = NULL,
  ymax = NULL,
  reflection = F,
  label.cex = 1,
  pt.color = "grey15",
  pt.cex = 0.2
)
}
\arguments{
\item{stats}{data frame of association statistics as read by \code{\link[=read.manhattan.stats]{read.manhattan.stats()}}}

\item{cutoff}{smallest P-value to render}

\item{highlights}{data frame of regions to highlight}

\item{highlight.color}{color to use for highlighted points}

\item{highlight.name}{label for highlighted points}

\item{print.stats}{boolean indicator to print stats on plot [default: TRUE]}

\item{echo.lambdas}{boolean indicator to print genomic control statistic to
stdout (for logging purposes) [default: FALSE]}

\item{legend}{boolean indicator to add legend [default: TRUE]}

\item{xmax}{maximum X-value to plot [default: include all points]}

\item{ymax}{maximum Y-value to plot [default: same as smallest P-value]}

\item{reflection}{boolean indicator to invert plot (useful for Miami plots)}

\item{label.cex}{scaling factor for label text}

\item{pt.color}{color for all points [default: "grey15"]}

\item{pt.cex}{scaling factor for all points [default: 0.2]}
}
\value{
None
}
\description{
Generate a Q-Q plot of association stats vs. uniform null
}
