% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.rects.for.highlight}
\alias{plot.rects.for.highlight}
\title{Plot simple rectangles for locus highlight}
\usage{
\method{plot}{rects.for.highlight}(
  xlefts,
  xrights,
  y0,
  col = blueblack,
  border = blueblack,
  panel.height = 0.2,
  rect.height.cex = 1,
  y.axis.title = NULL
)
}
\arguments{
\item{xlefts}{left-most position(s) for rectangle(s)}

\item{xrights}{right-most position(s) for rectangle(s)}

\item{col}{color(s) to fill rectangle(s)}

\item{border}{color(s) for rectangle outline(s)}

\item{panel.height}{relative height of panel [default: 0.2]}

\item{rect.height.cex}{relative vertical expansion scalar for rectangles
[default: 1]}

\item{y.axis.title}{title for Y axis [default: no title]}
}
\description{
Add panel of simple colored rectangles to locus highlight
}
