% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_plot.R
\name{plot.roc}
\alias{plot.roc}
\title{Plot ROC curve}
\usage{
\method{plot}{roc}(
  data,
  colors = NULL,
  nested.auc = TRUE,
  auc.text.colors = NULL,
  grid.col = bluewhite,
  diag.col = bluewhite,
  ax.tick = -0.025,
  parmar = c(2.5, 2.5, 0.75, 0.75)
)
}
\arguments{
\item{data}{list of outputs from \code{\link[=evaluate.score]{evaluate.score()}}}

\item{colors}{colors for each entry in \code{data} [default: use viridis()]}

\item{nested.auc}{boolean indicator to add nested AUC barplot [default: TRUE]}

\item{auc.text.colors}{colors for labels in nested AUC barplot}

\item{grid.col}{color of gridlines for plot [default: very light blue]}

\item{diag.col}{color of diagonal line [default: very light blue]}

\item{ax.tick}{\code{tck} value for X and Y axes [default: -0.025]}

\item{parmar}{values passed to \code{par(mar=...)}}
}
\description{
Plot receiver operating characteristic curves from a list of \code{\link[=evaluate.score]{evaluate.score()}} outputs
}
