% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{plot.seg.perms.multi}
\alias{plot.seg.perms.multi}
\title{Multi-panel segment permutation plot}
\usage{
\method{plot}{seg.perms.multi}(
  segs,
  gw.perms,
  lit.perms,
  union.perms,
  feature,
  measure,
  subset_to_regions = NULL,
  norm = F,
  n.bins = 100,
  min.bins = 10,
  x.title = NULL,
  xlims = NULL,
  xmin = NULL,
  xmax = NULL,
  max.x.ticks = 5,
  outer.axis.cex = 1,
  inner.axis.cex = 0.85,
  diamond.cex = 1.25,
  parmar = c(2.25, 6, 0.5, 0.5)
)
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{gw.perms}{list of permutation results for empirical segments}

\item{lit.perms}{list of permutation results for literature segments}

\item{union.perms}{list of permutation results for all combined segments}

\item{feature}{name of feature to evaluate}

\item{measure}{statistic to evaluate [default: mean]}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}

\item{norm}{normalize values before plotting [default: FALSE]}

\item{n.bins}{ideal number of bins for histograms [default: 100]}

\item{min.bins}{minimum number of bins [default: 10]}

\item{xlims}{limits for X axis}

\item{xmin}{minimum for X axis}

\item{xmax}{maximum for X axis}

\item{max.x.ticks}{maxmimum ticks for X axis}

\item{inner.axis.cex}{expansion of inner axis}

\item{diamond.cex}{expansion for observed values}

\item{parmar}{numeric vector of margins passed to \code{\link[=par]{par()}}}

\item{xtitle}{title of X axis}

\item{x.title.line}{line of X axis title}
}
\description{
Multi-panel plot of permutation results for segment subsets
}
\seealso{
\code{\link[=plot.seg.perms]{plot.seg.perms()}}
}
