% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{plot.seg.perms}
\alias{plot.seg.perms}
\title{Plot segment permutation results}
\usage{
\method{plot}{seg.perms}(
  segs,
  perms,
  feature,
  measure,
  norm = F,
  subset_to_regions = NULL,
  n.bins = 100,
  min.bins = 10,
  x.title = NULL,
  x.title.line = 1.2,
  x.title.cex = 1,
  xlims = NULL,
  xmin = NULL,
  xmax = NULL,
  diamond.pch = 23,
  diamond.cex = 1.25,
  parmar = c(2.25, 2, 0.5, 0.5)
)
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{perms}{data.table of permutation results}

\item{feature}{name of feature to evaluate}

\item{measure}{statistic to evaluate [default: mean]}

\item{norm}{normalize values before plotting [default: FALSE]}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}

\item{n.bins}{ideal number of bins for histograms [default: 100]}

\item{min.bins}{minimum number of bins [default: 10]}

\item{x.title.line}{line of X axis title}

\item{xlims}{limits for X axis}

\item{xmin}{minimum for X axis}

\item{xmax}{maximum for X axis}

\item{diamond.pch}{shape for observed values}

\item{diamond.cex}{expansion for observed values}

\item{parmar}{numeric vector of margins passed to \code{\link[=par]{par()}}}

\item{xtitle}{title of X axis}
}
\description{
Function to plot segment permutation test results
}
\seealso{
\code{\link[=plot.viohist]{plot.viohist()}}
}
