% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_plot.R
\name{plot.stratified.metric}
\alias{plot.stratified.metric}
\title{Plot metric stratified by gene scores}
\usage{
\method{plot}{stratified.metric}(strat.dat, ylims = NULL, y.title = NULL, parmar = c(1, 3, 0.5, 0.5))
}
\arguments{
\item{strat.dat}{list of values stratified as desired}

\item{ylims}{limits for Y axis}

\item{y.title}{title of Y axis}

\item{parmar}{values passed to \code{par(mar=...)}}
}
\description{
Plot a metric stratified by CNV type and low/high pHaplo & pTriplo
}
