% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{prc}
\alias{prc}
\title{Compute gene score precision-recall statistics}
\usage{
prc(stats, score, true.genes, false.genes, steps = seq(1, 0, -0.001))
}
\arguments{
\item{stats}{data.frame of scores for all genes}

\item{score}{score to evaluate; must be a column of \code{stats}}

\item{true.genes}{vector of gene symbols to consider true positives}

\item{false.genes}{vector of gene symbols to consider true negatives}

\item{steps}{numeric vector of score cutoffs to evaluate [default: from 1 to 0 in steps of 0.001]}
}
\value{
data.frame of P-R statistics at each score cutoff
}
\description{
Compute precision-recall of a single score vs. predefined true/false genes
}
