% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{prune.hpo.jaccard.matrix}
\alias{prune.hpo.jaccard.matrix}
\title{Prune HPO Jaccard matrix}
\usage{
prune.hpo.jaccard.matrix(
  hpo.jac,
  sample.sizes,
  max.jac = 0.5,
  ignore.hpos = c("HP:0000118", "UNKNOWN")
)
}
\arguments{
\item{hpo.jac}{square matrix of Jaccard indexes for all HPOs}

\item{sample.sizes}{named vector of sample sizes for each HPO}

\item{max.jac}{maximum Jaccard index to retain [default: 0.5]}

\item{ignore.hpos}{vector of HPOs to ignore [default: "HP:0000118" & "UNKNOWN"]}
}
\value{
data.frame filtered on \code{max.jac}
}
\description{
Prune a matrix of Jaccard indexes for all HPO pairs based on maximum Jaccard index
}
\details{
Pairs will be pruned while retaining the HPO with the larger sample size
}
\seealso{
\link{load.hpo.jaccard.matrix}
}
