% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{reorganize.perm.plots}
\alias{reorganize.perm.plots}
\title{Reorganize permutation plots}
\usage{
reorganize.perm.plots(parent.dir)
}
\arguments{
\item{parent.dir}{Parent directory containing all permutation plots}
}
\description{
Reorganize all key permutation plots for a single subset of segments
}
