% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{robust.lm}
\alias{robust.lm}
\title{Fit a robust linear regression with confidence interval}
\usage{
robust.lm(x, y, conf = 0.95)
}
\arguments{
\item{x}{independent variable vector}

\item{y}{dependent variable vector}

\item{conf}{width of confidence interval [default: 0.95]}
}
\value{
list of two objects:
\enumerate{
\item \verb{$fit} : fitted model
\item \verb{$ci}: data frame with bounds of confidence interval
}
}
\description{
Fits an outlier-robust version of univariate linear regression with confidence interval
}
\details{
Effectively wraps \code{MASS:rlm()} and handles data reformatting as needed
}
