% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{saddlepoint.adj}
\alias{saddlepoint.adj}
\title{Z-score saddlepoint approximation}
\usage{
saddlepoint.adj(
  zscores,
  winsorize = 1,
  winsorize.left.tail = F,
  phred = T,
 
    min.p = 0.000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
)
}
\arguments{
\item{zscores}{numeric vector of Z-scores}

\item{winsorize}{width of quantile intervals to Winsorize [default: 1, i.e., no Winsorization]}

\item{winsorize.left.tail}{boolean indicator to apply Winsorization to left tail of distribution [default: FALSE]}

\item{phred}{boolean indicator of whether to -log10-scale adjusted P-values}

\item{min.p}{smallest P-value to report (P-values smaller than this will be rounded up) [default: 1e-300]}
}
\value{
numeric vector of adjusted P-values
}
\description{
Apply saddlepoint approximation to vector of Z-scores to generate adjusted P-values
}
