% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{segs.simple.vioswarm}
\alias{segs.simple.vioswarm}
\title{Simple segment violin/swarmplot}
\usage{
segs.simple.vioswarm(
  segs,
  y,
  subset_to_regions = NULL,
  add.y.axis = T,
  ytitle = NULL,
  ytitle.line = 1.75,
  add.pvalue = FALSE,
  stat.test = "wilcoxon",
  alternative = "two.sided",
  pt.cex = 1,
  parmar = c(1.3, 3, 0.3, 0.3)
)
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{y}{numeric vector for Y axis}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}

\item{add.y.axis}{add Y axis to plot [default: TRUE]}

\item{ytitle}{title of Y axis}

\item{add.pvalue}{print P-value on plot [default: FALSE]}

\item{stat.test}{statistical test to apply to groups [default: \code{wilcoxon}]}

\item{alternative}{alternative hypothesis for \code{stat.test} [default: \code{two.sided}]}

\item{pt.cex}{expansion factor for points[default: 1]}

\item{parmar}{numeric vector of margins passed to \code{\link[=par]{par()}}}

\item{y.title.line}{line of Y axis title}
}
\description{
Simpler generic vioplot/swarmplot hybrid for plotting values split by DEL & DUP
}
\seealso{
\code{\link[=segs.swarm]{segs.swarm()}}
}
