% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{segs.swarm}
\alias{segs.swarm}
\title{Segment swarm/boxplot}
\usage{
segs.swarm(
  segs,
  x.bool,
  y,
  cnv.split = TRUE,
  ylims = NULL,
  subset_to_regions = NULL,
  add.pvalue = FALSE,
  stat.test = "wilcoxon",
  alternative = "two.sided",
  print.fold = FALSE,
  xtitle = NULL,
  x.labs = c("FALSE", "TRUE"),
  add.y.axis = TRUE,
  ytitle = NULL,
  y.title.line = 1.75,
  y.at = NULL,
  y.labs = NULL,
  y.labs.at = NULL,
  parse.y.labs = FALSE,
  violin = FALSE,
  pt.cex = 1,
  parmar = c(2.3, 3, 0.5, 0.5)
)
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{x.bool}{boolean vector for subgrouping values on X axis}

\item{y}{numeric vector for Y axis}

\item{cnv.split}{split by CNV type [default: TRUE]}

\item{ylims}{Y axis limits}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}

\item{add.pvalue}{print P-value on plot [default: FALSE]}

\item{stat.test}{statistical test to apply to groups [default: \code{wilcoxon}]}

\item{alternative}{alternative hypothesis for \code{stat.test} [default: \code{two.sided}]}

\item{print.fold}{print fold-enrichment to stdout [default: FALSE]}

\item{xtitle}{title of X axis}

\item{x.labs}{character vector for labels on X axis}

\item{add.y.axis}{add Y axis to plot [default: TRUE]}

\item{ytitle}{title of Y axis}

\item{y.title.line}{line of Y axis title}

\item{y.at}{tick positions for Y axis}

\item{y.labs}{labels for Y axis ticks}

\item{y.labs.at}{if different from \code{y.at}, specify locations of \code{y.labs}}

\item{parse.y.labs}{format Y axis labels with \code{\link[=parse]{parse()}} [default: FALSE]}

\item{violin}{plot violins instead of boxes [default: FALSE]}

\item{pt.cex}{expansion factor for points[default: 1]}

\item{parmar}{numeric vector of margins passed to \code{\link[=par]{par()}}}
}
\description{
Generic swarm/boxplot function for large segment analyses
}
