% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_plot.R
\name{simple.legend}
\alias{simple.legend}
\title{Plot simple legend}
\usage{
simple.legend(labels, colors)
}
\arguments{
\item{labels}{vector of legend labels}

\item{colors}{vector of legend colors}
}
\description{
Plot simple color legend
}
