% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{standardize.effect.size}
\alias{standardize.effect.size}
\title{Standardize effect sizes}
\usage{
standardize.effect.size(or, ci.lower, ci.upper, alpha = 0.05)
}
\arguments{
\item{or}{point estimate of odds ratio}

\item{ci.lower}{lower bound of odds ratio confidence interval}

\item{ci.upper}{upper bound of odds ratio confidence interval}

\item{alpha}{alpha value corresponding to confidence interval [default: 0.05]}
}
\value{
numeric vector of standardized effect sizes reported as log-odds ratio
divided by standard deviation
}
\description{
Standardize effect sizes generated by single-cohort association tests
}
