% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_assoc_utils.R
\name{summarize.finemapping.categories}
\alias{summarize.finemapping.categories}
\title{Summarize categories of fine-mapped genes}
\usage{
summarize.finemapping.categories(g.cats, cnv = "CNV")
}
\arguments{
\item{g.cats}{categories of genes split by PIP & rank as produced by \code{\link[=categorize.genes]{categorize.genes()}}}

\item{cnv}{CNV type to evaluate [default: "CNV"]}
}
\description{
Summarize fine-mapping categories based on PIP & rank in each credible set
}
