% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_plot.R
\name{superimposed.barplot}
\alias{superimposed.barplot}
\title{Barplot inset for score performance}
\usage{
superimposed.barplot(
  values,
  colors,
  xleft,
  xright,
  ybottom,
  ytop,
  min.value = NULL,
  max.value = NULL,
  title = NULL,
  add.labels = TRUE,
  lab.cex = 0.6,
  lab.colors = NULL,
  buffer = 0.025
)
}
\arguments{
\item{values}{values to be plotted}

\item{colors}{colors for each bar}

\item{xleft}{leftmost position of inset on X axis}

\item{xright}{rightmost position of inset on X axis}

\item{ybottom}{bottommost position of inset on Y axis}

\item{ytop}{topmost position of inset on Y axis}

\item{min.value}{minimum value to plot}

\item{max.value}{maximum value to plot}

\item{title}{title of plot}

\item{add.labels}{boolean indicator whether to label bars [default: TRUE]}

\item{lab.cex}{cex value for bar labels [default: 0.6]}

\item{lab.colors}{colors for bar labels}

\item{buffer}{relative buffer between bars}
}
\description{
Superimposed barplot of one statistic per category from \code{\link[=evaluate.scores]{evaluate.scores()}}
}
\seealso{
\code{\link[=evaluate.score]{evaluate.score()}}
}
