% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{calc.or}
\alias{calc.or}
\title{Calculate CNV odds ratio}
\usage{
calc.or(
  control_ref,
  control_alt,
  case_ref,
  case_alt,
  adj = 0.5,
  return.confint = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{control_ref}{number of control non-carriers}

\item{control_alt}{number of control CNV carriers}

\item{case_ref}{number of case non-carriers}

\item{case_alt}{number of case CNV carriers}

\item{adj}{fixed value to use for continuity correction [default: 0.5]}

\item{return.confint}{return confidence interval [default: false]}

\item{alpha}{alpha value corresponding to \code{return.confint} [default: 0.05]}
}
\value{
numeric vector of odds ratio, optionally with lower and upper 95\% confidence bounds
}
\description{
Compute basic OR of CNV carriers with adjustment for zero-inflation
}
\details{
implements simple continuity correction with fixed value added to
all cells of contingency table
}
