% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_score_utils.R
\name{classify.genes.by.score}
\alias{classify.genes.by.score}
\title{Partition genes by dosage sensitivity}
\usage{
classify.genes.by.score(scores, hc.cutoff = 0.9, lc.cutoff = 0.5)
}
\arguments{
\item{scores}{data.frame of gene scores as loaded by \code{\link[=load.scores]{load.scores()}}}

\item{hc.cutoff}{minimum score for high-confidence designation [default: 0.9]}

\item{lc.cutoff}{minimum score for low-confidence designatnion [default: 0.5]}
}
\value{
list of gene symbols for each group
}
\description{
Partition genes into subgroups based on dosage sensitivity scores
}
\seealso{
\code{\link[=load.scores]{load.scores()}}
}
