% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{combine.single.cohort.assoc.stats}
\alias{combine.single.cohort.assoc.stats}
\title{Merge single-cohort association stats}
\usage{
combine.single.cohort.assoc.stats(stats.list, cond.excl.in = NULL)
}
\arguments{
\item{stats.list}{list of single-cohort association stats}

\item{cond.excl.in}{path to BED file of cohorts to be excluded on locus-specific basis}
}
\value{
data frame of association stats for all cohorts
}
\description{
Merges a list of association statistics generated for multiple independent cohorts
}
\seealso{
\code{\link[=read.assoc.stats.single]{read.assoc.stats.single()}}
}
