% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{estimate.cohort.inflation}
\alias{estimate.cohort.inflation}
\title{Estimate cohort inflation terms}
\usage{
estimate.cohort.inflation(stats.merged, cohorts)
}
\arguments{
\item{stats.merged}{association statistics merged across cohorts as read by \code{\link[=combine.single.cohort.assoc.stats]{combine.single.cohort.assoc.stats()}}}

\item{cohorts}{list of cohorts to evaluate}
}
\value{
named vector with mean log odds ratio for each cohort
}
\description{
Estimates mean genome-wide effect size inflation per cohort
}
\seealso{
\code{\link[=combine.single.cohort.assoc.stats]{combine.single.cohort.assoc.stats()}}
}
