% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get.hpo.color}
\alias{get.hpo.color}
\title{Get HPO-based color code}
\usage{
get.hpo.color(hpo, color.by = "neuro")
}
\arguments{
\item{hpo}{HPO code}

\item{color.by}{indicator to color by 'neuro' or 'severity'}
}
\value{
color code
}
\description{
Return hex color code for phenotype by HPO
}
