% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load.gene.feature.metadata}
\alias{load.gene.feature.metadata}
\title{Load gene feature metadata}
\usage{
load.gene.feature.metadata(feature.metadata.in)
}
\arguments{
\item{feature.metadata.in}{path to gene feature metadata .tsv}
}
\value{
data.frame of contents from \code{feature.metadata.in}
}
\description{
Load table of gene feature metadata (including plain-text descriptions)
}
