% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_utils.R
\name{load.sumstats.for.region}
\alias{load.sumstats.for.region}
\title{Load sumstats from BED}
\usage{
load.sumstats.for.region(bedpath, region, rstudio.local = FALSE)
}
\arguments{
\item{bedpath}{path to .bed file with summary statistics}

\item{region}{coordinates of region to be extracted}

\item{rstudio.local}{boolean to indicate local Rstudio environment [default: FALSE]}
}
\value{
data.frame
}
\description{
Extract meta-analysis summary statistics from a single BED
}
\details{
\code{region} coordinates must be a \code{tabix}-compatbile string
}
