% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{make.meta.lookup.table}
\alias{make.meta.lookup.table}
\title{Create meta-analysis lookup table}
\usage{
make.meta.lookup.table(stats.merged, cohorts, model, empirical.continuity = T)
}
\arguments{
\item{stats.merged}{association statistics for all cohorts (see \code{\link[=combine.single.cohort.assoc.stats]{combine.single.cohort.assoc.stats()}})}

\item{cohorts}{vector of all cohort names}

\item{model}{specify meta-analysis model to use (see \code{Details})}

\item{empirical.continuity}{boolean indicator to apply Sweeting empirical
continuity correction [default: TRUE]}
}
\value{
data frame of all unique case & control CNV counts
}
\description{
Make meta-analysis lookup table to shorten time required to run full meta-analysis
}
\details{
\code{model} accepts several character inputs, including:
\itemize{
\item \code{fe} : fixed-effects model implemented by \code{\link[metafor:rma.uni]{metafor::rma.uni()}} [default]
\item \code{re} : random-effects model implemented by \code{\link[metafor:rma.uni]{metafor::rma.uni()}}
\item \code{mh} : Mantel-Haenszel model implemented by \code{\link[metafor:rma.mh]{metafor::rma.mh()}}
}
}
