% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{perm.summary}
\alias{perm.summary}
\title{Summarize segment permutation results}
\usage{
perm.summary(perms.orig, feature, measure = "mean", subset_to_regions = NULL)
}
\arguments{
\item{perms.orig}{data.table of permutation results}

\item{feature}{name of feature to evaluate}

\item{measure}{statistic to evaluate [default: mean]}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}
}
\value{
matrix of statistic per permutation split by all, deletions, and duplications
}
\description{
Summarize permutation results across all permutations for a single feature
}
\details{
Valid options for \code{measure} are mean, median, sum, and frac.any
}
