% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_plot.R
\name{plot.idio.stick}
\alias{plot.idio.stick}
\title{Add idiogram line}
\usage{
\method{plot}{idio.stick}(
  genome.in,
  chrom,
  y.at,
  start = NA,
  end = NA,
  tick.height = -0.03,
  wex = 0.925,
  y.label = TRUE
)
}
\arguments{
\item{genome.in}{BEDTools-style .genome file}

\item{chrom}{chromosome to be plotted}

\item{y.at}{where the idiogram should be drawn, in Y coordinates}

\item{start}{left-most coordinate of highlight region [default: no highlight]}

\item{end}{right-most coordinate of highlight region [default: no highlight]}

\item{tick.height}{relative height of idiogram ticks [default: -0.03]}

\item{wex}{relative width expansion factor [default: 0.925]}

\item{y.label}{boolean indicator to add chromosome label to the Y axis [default: TRUE]}
}
\description{
Add stick figure for idiogram in locus highlights
}
