% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_qq.R
\name{plot.manhattan}
\alias{plot.manhattan}
\title{Manhattan plot}
\usage{
\method{plot}{manhattan}(
  df,
  cutoff = 0.00000001,
  highlights = NULL,
  highlight.color = "#4EE69A",
  highlight.name = "Highlighted Loci",
  mark.highlights = FALSE,
  colors = c("gray15", "gray70"),
  cutoff.color = "#D01C8B",
  lab.prefix = NULL,
  ymax = NULL,
  reflection = F,
  pt.cex = 0.2,
  label.cex = 1,
  y.ax.cex = 0.75
)
}
\arguments{
\item{df}{data frame of association statistics as read by \code{\link[=read.manhattan.stats]{read.manhattan.stats()}}}

\item{cutoff}{smallest P-value to render [default: 10^-8]}

\item{highlights}{data frame of regions to highlight}

\item{highlight.color}{color to use for highlighted points}

\item{highlight.name}{label for highlighted points}

\item{mark.highlights}{boolean indicated to mark highlighted regions with small arrows [default: FALSE]}

\item{colors}{vector of colors to assign sequentially to all chromosomes}

\item{cutoff.color}{color to use for P-value cutoff line}

\item{lab.prefix}{prefix to prepend to various labels}

\item{ymax}{maximum Y-value to plot [default: same as smallest P-value]}

\item{reflection}{boolean indicator to invert plot (useful for Miami plots)}

\item{pt.cex}{scaling factor for points [default: 0.2]}

\item{label.cex}{scaling factor for label text}

\item{y.ax.cex}{scaling factor for labels on Y-axis [default: 0.75]}
}
\value{
None
}
\description{
Generate a Manhattan plot from association statistics
}
