% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{saddlepoint.adj}
\alias{saddlepoint.adj}
\title{Z-score saddlepoint approximation}
\usage{
saddlepoint.adj(zscores, phred = T)
}
\arguments{
\item{zscores}{numeric vector of Z-scores}

\item{phred}{boolean indicator of whether to -log10-scale adjusted P-values}
}
\value{
numeric vector of adjusted P-values
}
\description{
Apply saddlepoint approximation to vector of Z-scores to generate adjusted P-values
}
