% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{saddlepoint.adj}
\alias{saddlepoint.adj}
\title{Z-score saddlepoint approximation}
\usage{
saddlepoint.adj(
  zscores,
  winsorize = 1,
  winsorize.left.tail = F,
  mirror = F,
  phred = T
)
}
\arguments{
\item{zscores}{numeric vector of Z-scores}

\item{winsorize}{width of quantile intervals to Winsorize [default: 1, i.e., no Winsorization]}

\item{winsorize.left.tail}{boolean indicator to apply Winsorization to left tail of distribution [default: FALSE]}

\item{mirror}{mirror bottom 50\% of Z-scores [default: FALSE]}

\item{phred}{boolean indicator of whether to -log10-scale adjusted P-values}
}
\value{
data.frame with two columns:
\itemize{
\item \verb{$zscores} for corrected Z-scores
\item \verb{$pvalues} for P-values corresponding to corrected Z-scores
}
}
\description{
Apply saddlepoint approximation to vector of Z-scores to generate adjusted P-values
}
