% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_plot.R
\name{segs.scatter}
\alias{segs.scatter}
\title{Generic segment scatterplot}
\usage{
segs.scatter(
  segs,
  x,
  y,
  subset_to_regions = NULL,
  xlims = NULL,
  ylims = NULL,
  add.lm = T,
  add.cor = add.lm,
  pt.cex = 1,
  blue.bg = TRUE,
  horiz.lines.at = NULL,
  horiz.lines.lty = 1,
  horiz.lines.color = NULL,
  abline.a = NULL,
  abline.b = NULL,
  abline.lty = 1,
  xtitle = NULL,
  x.title.line = 1.75,
  x.at = NULL,
  x.labs = NULL,
  x.labs.at = NULL,
  parse.x.labs = FALSE,
  x.labs.line = -0.6,
  max.x.ticks = 6,
  x.title.cex = 1,
  ytitle = NULL,
  y.title.line = 1.75,
  y.at = NULL,
  y.labs = NULL,
  y.labs.at = NULL,
  parse.y.labs = FALSE,
  y.labs.line = -0.6,
  max.y.ticks = 6,
  y.title.cex = 1,
  parmar = c(3, 3, 0.8, 0.8)
)
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{x}{numeric vector for X axis}

\item{y}{numeric vector for Y axis}

\item{subset_to_regions}{vector of region IDs to include [default: include all regions]}

\item{xlims}{X axis limits}

\item{ylims}{Y axis limits}

\item{add.lm}{add linear trendline [default: TRUE]}

\item{add.cor}{add correlation coefficients & P-values for linear model [default: TRUE]}

\item{pt.cex}{expansion factor for points [default: 1]}

\item{blue.bg}{add light blue background [default: TRUE]}

\item{horiz.lines.at}{numeric vector indicating where to draw horizontal lines [default: no lines]}

\item{horiz.lines.lty}{lty parameter for \code{horiz.lines.at} [default: 1]}

\item{horiz.lines.color}{colors for \code{horiz.lines.at}}

\item{abline.a}{\code{a} value for embedded \code{\link[=abline]{abline()}} call}

\item{abline.b}{\code{b} value for embedded \code{\link[=abline]{abline()}} call}

\item{abline.lty}{lty parameter for embedded \code{\link[=abline]{abline()}} call}

\item{xtitle}{title of X axis}

\item{x.title.line}{line of X axis title}

\item{x.at}{tick positions for X axis}

\item{x.labs}{labels for X axis ticks}

\item{x.labs.at}{if different from \code{x.at}, specify locations of \code{x.labs}}

\item{parse.x.labs}{format X axis labels with \code{\link[=parse]{parse()}} [default: FALSE]}

\item{x.labs.line}{line for X axis labels [default: -0.6]}

\item{max.x.ticks}{maximum number of X axis ticks to display [default: 6]}

\item{x.title.cex}{expansion factor for X axis title [default: 1]}

\item{ytitle}{title of Y axis}

\item{y.title.line}{line of Y axis title}

\item{y.at}{tick positions for Y axis}

\item{y.labs}{labels for Y axis ticks}

\item{y.labs.at}{if different from \code{y.at}, specify locations of \code{y.labs}}

\item{parse.y.labs}{format Y axis labels with \code{\link[=parse]{parse()}} [default: FALSE]}

\item{y.labs.line}{line for Y axis labels [default: -0.6]}

\item{max.y.ticks}{maximum number of Y axis ticks to display [default: 6]}

\item{y.title.cex}{expansion factor for Y axis title [default: 1]}

\item{parmar}{numeric vector of margins passed to \code{\link[=par]{par()}}}
}
\description{
Generic segment scatterplot function
}
\seealso{
\code{\link[=load.segment.table]{load.segment.table()}}
}
