% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_seg_utils.R
\name{split.regions.by.effect.size}
\alias{split.regions.by.effect.size}
\title{Split segments by effect size}
\usage{
\method{split}{regions.by.effect.size}(segs, quantiles = 2, effect.size.colname = "meta_best_lnor")
}
\arguments{
\item{segs}{segment dataframe (imported with \code{\link[=load.segment.table]{load.segment.table()}})}

\item{quantiles}{number of quantiles to divide [default: 2]}

\item{effect.size.colname}{name of column to reference for effect sizes
[default: \code{meta_best_lnor}]}
}
\value{
list of length \code{quantiles} with region IDs for each quantile
}
\description{
Split segments into groups based on effect size quantiles
}
\details{
uses values from \code{effect.size.colname} \(see above\) while
matching on CNV type to compute quantile cutoffs
}
